pyproject.toml
[project]
name = "classic-books-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.0",
    "bs4>=0.0.2",
    "certifi>=2025.10.5",
    "fastapi>=0.119.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "setuptools>=80.9.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
backend = [
    "onnxruntime-gpu>=1.24.3",
    "uvicorn>=0.38.0",
]
frontend = [
    "httpx>=0.28.1",
    "streamlit>=1.50.0",
]
ml = [
    "datasets>=4.6.1",
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "mlflow>=3.10.0",
    "nbconvert>=7.17.0",
    "nbformat>=5.10.4",
    "nltk>=3.9.3",
    "onnx>=1.20.1",
    "optuna>=4.7.0",
    "peft>=0.18.1",
    "plotly>=6.6.0",
    "pymorphy2>=0.9.1",
    "torch>=2.8.0",
    "torch-onnx>=0.1.25",
    "transformers>=5.0.0",
]
